requirements.txt
# Существующие зависимости
altgraph==0.17.4
et_xmlfile==2.0.0
lxml==6.0.1
macholib==1.16.3
openpyxl==3.1.5
packaging==25.0
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PySide6==6.9.2
PySide6_Addons==6.9.2
PySide6_Essentials==6.9.2
python-docx==1.2.0
setuptools==80.9.0
shiboken6==6.9.2
typing_extensions==4.15.0
pre-commit==4.3.0
black==25.9.0
flake8==7.3.0
isort==7.0.0
pytest==8.4.2
